Use --query_file for bazel query to avoid argument length errors
Extract shared bazel query utility (tools/ci/bazel_query.py) and update
both ci_decide.py and bazel_diff.py to use it. The --query_file flag
avoids "Argument list too long" errors when querying large target sets.
